Add test coverage for new Backhole funcs
